Make less decide requests when enriching events
When `sendFeatureFlags` is `true`, the client makes a `Decide` requests to get the latest feature flags in order to enrich the captured event with the current feature flag values.

However, when processing a batch, it doesn't make sense to make a decide request multiple times for the same `distinctId`. This commit adds an in-memory cache that lasts for the duration of the batch.